% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SFAmixWrapper.R
\name{gen_SFAmix_data}
\alias{gen_SFAmix_data}
\title{Simulate matrix with dimension of 500 x 200, number of factors is set to 15, where 10 of them being sparse. The sparse loading matrix cotains mostly zeros, and random blocks of nonzero values generated from N(0,std). The dense loading matrix is generated from N(0,std), the factor matrix and the error matrix are generated from N(0,1).}
\usage{
gen_SFAmix_data(std = 1)
}
\arguments{
\item{std}{standard deviation for the normal distribution}
}
\value{
a list containing the following

lams: the sparse loading matrix

lamd: the dense loading matrix

ex: the factors matrix

y: the y matrix calculated as y = lam * ex + err
}
\description{
Simulate matrix with dimension of 500 x 200, number of factors is set to 15, where 10 of them being sparse. The sparse loading matrix cotains mostly zeros, and random blocks of nonzero values generated from N(0,std). The dense loading matrix is generated from N(0,std), the factor matrix and the error matrix are generated from N(0,1).
}
